added internal directory and moved core package inside it
